QOLOE-120 Enhancements (#171)
