+++ src/components/bs5/header/header.scss
@@ -46,7 +46,7 @@
 				color: var(--#{$prefix}header_skip-link_focus_text_color);
 				background-color: var(--#{$prefix}header_skip-link_focus_bg_color);
 				outline: 3px solid var(--#{$prefix}header_skip-link_focus_outline_color);
-				z-index: 600;
+				z-index: 999;
 			}
 		}
 	}
@@ -207,20 +207,15 @@
 				font-size: .875rem;
 				line-height: 1.42;
 				padding-left: .75rem;
+				margin: 0 .5rem;
 				align-items: center;
 				color: var(--#{$prefix}header__cta-wrapper__cta-link_text_color);
-
-				&:hover {
-					.qld__header__cta-link-icon {
-						color: var(--#{$prefix}header__cta-wrapper__cta-link-icon_hover_color);
-					}
-				}
 			}
 
 			.qld__header__cta-link-icon {
-				height: 1.25rem;
-				width: 1.25rem;
-				margin-right: .5rem;
+				height: 1rem;
+				width: 1rem;
+				margin-left: .5rem;
 				text-align: center;
 				display: inline-block;
 				color: var(--#{$prefix}header__cta-wrapper__cta-link-icon_color);
@@ -475,153 +470,9 @@
 		form {
 			display: flex;
 			position: relative;
-
-			::placeholder {
-				color: var(--site-search-input-color);
-			}
-
-			//Search dropdown
-			.suggestions-dropdown {
-				position: absolute;
-				display: none;
-				left: 0;	
-				top: 100%;  
-				width: 100%;			
-				z-index: 99; 
-				border-radius: .5rem;
-				background: var(--#{$prefix}header_site-search-suggestions-bg);
-				box-shadow: var(--#{$prefix}header__pre-header_boxshadow);
-				border-bottom: solid .25rem var(--#{$prefix}header_site-search-suggestions-hover__border_color);
-
-				.suggestions-category {
-					padding: 0 1rem;
-				}
-
-				hr {
-					margin: 0;
-				}
-
-				// Featured search result styles
-				.feature {
-					background-color: var(--#{$prefix}header_site-search-suggestions-feature_bg);
-					strong {
-						color: var(--#{$prefix}header_site-search-suggestions-feature_text-color);
-					}
-					ul {
-						li {
-							&:hover {
-								background-color: var(--#{$prefix}header_site-search-suggestions-feature_hover);
-							}
-							a {
-								color: var(--#{$prefix}header_site-search-suggestions-feature_text-color);
-							}
-						}
-					}				
-				}
-
-				// Show when active
-				&.show {
-					display: block; 
-				}
-
-				&__group {
-					padding: 0;
-
-					a, strong {
-						color: var(--site-search-input-color);
-						font-weight: 400;
-						font-size: 1rem;
-						text-decoration: none; 
-					}
-
-					strong {
-						font-weight: 600;
-					}
-
-					ul {
-						padding: 0;
-
-						li {
-							min-height: 3rem;
-							list-style: none;
-							cursor: pointer;
-
-							a {
-								vertical-align: middle;
-								vertical-align: -webkit-baseline-middle;
-								&:hover {
-									text-decoration: underline; 
-								}
-							}
-
-							&:hover {
-								background-color: var(--#{$prefix}header_site-search-suggestions-hover);
-								margin: 0 -1rem;
-								padding: 0 1rem;
-
-								a {
-									text-decoration: underline;
-								}
-
-							}
-						}
-					}
-				}
-			}
-
-			input[type="text"] {
-				box-shadow: none;
-				border: none;
-				padding-left: 3rem;
-				width: 20.62rem;
-				color: var(--site-search-input-color);
-				background: var(--site-search-bg);
-				border-radius: 0.25rem 0 0 0;
-				border-bottom: 2px solid var(--site-search-border-color);
-
-				@include media-breakpoint-down(lg) {
-					border: 2px solid var(--site-search-border-color);
-					width: 100%;
-				}
-
-				&:focus {
-					outline: 2px solid var(--site-search-input-focus-color);
-					outline-offset: 2px;
-				}
-
-				&.border-full {
-					border: 2px solid var(--site-search-border-color);
-					border-right: 0;
-					border-radius: 0.25rem 0 0 0.25rem;
-				}
-
-			}
-
-			input[type="submit"] {
-				border: none;
-				box-shadow: none;
-				border-radius: 0 0.25rem 0.25rem 0;
-				background: var(--site-search-btn-bg);
-				color: var(--site-search-btn-color);
-			}
-
-			//Icon
-			&:before {
-				content: "";
-				position: absolute;
-				top: 0;
-				left: 0;
-				bottom: 2px;
-				width: 3rem;
-				mask-position: center center;
-				mask-repeat: no-repeat;
-				mask-image: var(--icon-search);
-				background-color: var(--site-search-icon-color);
-			}
 		}
-
+		
 		@include media-breakpoint-down(lg) {
-
 			// Mobile search open class
 			&--open {
 				z-index: 1;
@@ -633,10 +484,10 @@
 				margin-top: 1rem;
 				background-color: var(--#{$prefix}header_bg);
 			}
-
 			&--closed {
 				display: none;
 			}
 		}
 	}
+
 }
