+++ src/components/bs5/navbar/navbar.hbs
@@ -94,7 +94,7 @@
                             {{title}}
                         </a>
 
-                        <button class="nav-link dropdown-toggle second-element" id="menuDropdown" role="button"
+                        <button class="dropdown-toggle second-element" id="menuDropdown" role="button"
                             aria-controls="" data-bs-toggle="dropdown" aria-expanded="false" aria-selected="false"
                             aria-label="Toggle navigation">
                             <svg class="toggle_icon">
