+++ src/components/bs5/searchInput/searchInput.hbs
@@ -2,15 +2,21 @@
     QGDS Component: Search input 
 -->
 <div class="container {{variantClass}}">
-    
+
     <div class="qld-search-input {{customClass}}">
 
-        <input id="{{ inputID }}" name=" {{ inputName }}" class="form-control" type="text" placeholder="{{placeholder}}" aria-label="{{ ariaLabel }}" />
-    
+        <input id="{{ inputID }}" name=" {{ inputName }}" class="form-control" type="text" placeholder="{{placeholder}}"
+            aria-label="{{ ariaLabel }}" {{#each tags}} data-{{@key}}="{{this}}" {{/each}} />
+
         <button class="btn btn-primary" type="{{ buttonType }}" id="{{ buttonID }}">
             <span class="btn-icon"></span>
             <span class="btn-label">{{ buttonLabel }}</span>
         </button>
 
+        {{#if suggestions}}
+        <ul class="suggestions suggestions__group"></ul>
+        {{/if}}
+
     </div>
+
 </div>
